backend/requirements.txt
fastapi
sqlalchemy
pydantic
pyjwt
python-dotenv
pymysql
bcrypt